completed transaction route testing
